[WIP] Prefix-based endpoint selection
** NOTE **

No (unit) tests have been added yet for the new feature.  This can be
done when the core code has been accepted by the maintainers.

To ease testing, the commit contains verbose logging that should
be removed for production use.

This change deprecates the --registry-mirrors and --insecure-registries
flags and the counterpart options in the daemon.json config.

** PREFIXES AND PREFIX MATCHING **

A prefix is used to select a user-specified set of registry endpoints
based on an *image reference*.  When looking up registry endpoints
during a push or pull operation, the daemon checks if a specified
prefix is a string prefix of the current image reference.

For instance, the prefix "docker.io/library/" is a string prefix of
"docker.io/library/alpine:latest" and would match during a
`docker pull alpine`, `docker pull alpine:latest` and
`docker pull docker.io/library/alpine:latest`.  If multiple prefixes
match, the longest one will be selected.

Please note that prefixes match only at path/namespace boundaries, and
do not match an image name (e.g., "docker.io/library/alpine"), an image
tag (e.g., "docker.io/library/alpine:TAG") or at digest boundaries
(e.g., "docker.io/library/alpine@DIGEST").  To avoid internal magic and
unintended surprises for users, an error is returned if a specified
prefix doesn't end with a '/'.

** PUSH/PULL SEMANTICS **

When pushing, at most one specified endpoint will be used.  When
pulling, all specified endpoints will be used in the order of the
prefix matching appearance.

Consider we have specified the following push and pull endpoints:

push = [A,B,C]
pull = [A,D,E,F]

If the specified image of a push request matches the prefix of push
endpoints A and C, only the first matching registry endpoint (i.e.,
A) will be used; all other prefix-matching endpoints will be ignored.

If the specified image of a pull request matches the prefix of all pull
endpoints, the daemon will attempt to pull from the endpoints in the
order of matching appearance (i.e., A, D, E, F) and return the result of
the first successful pull or return an error if all attempts failed.

** CREDENTIAL HANDLING **

Credential handling does not adhere to the prefix semantics of this
change.  Any request to a registry that requires authentication will
relate to original request and not to any potential registry matching
the prefix.

A minimal example:

prefix  =  docker.io/library
address = secure.internal.registry.com:5000

If `secure.internal.registry.com:5000` requires authentication, requests
to `docker.io/library` will fail.  For instance, a `docker pull
docker.io/library/alpine:latest` will fail as the credentials for
docker.io/library would be used (i.e., the ones matching the original
request), which are likely to be different to the ones of
`secure.internal.registry.com:5000`.

A short-term fix is to downstream patch the daemon to allow configuring
the daemon.json config to include optional user/password for each
specified push/pull endpoint.  This has been tested succesfully but was
rejected due to security concerns as it implies storing credentials in
clear text - generally a bad practice and undesirable.

On the long term, the credential handling needs to be extended in the
clients (Docker client, docker-shim, CRI, etc.) and in the container
engines (Docker, containerd, CRI-O) simultaneously.  How this can be
achieved (e.g., clients send all credentials, a subset, etc.) remains
unclear and is subject to future work.  The K8s folks must be involved
as well.

** OPEN ISSUES **

As @thaJeztah has pointed out in #34319 (comment), the daemon is mixing
hostnames and images pointing to the Docker hub in a way that `docker
pull registry-1.docker.io/library/foo` yields the same image as `docker
pull docker.io/library/foo` and `docker pull foo`.  Is that something we
should "fix" to allow only "docker.io" as a valid namespace and not
translate "registry-1.docker.io" to "docker.io"?

Another question: Should the default registry endpoint (i.e.,
registry-1.docker.io) be in the push/pull endpoints for all prefixes
starting with "docker.io/" by default?  The reasoning behind is that it
would be more user-friendly and align with the behaviour of the
deprecated --registry-mirrors option.

Signed-off-by: Flavio   Castelli <fcastelli@suse.com>
Signed-off-by: Valentin Rothberg <vrothberg@suse.com>